I'm not a hundred percent sure, when this happens, but under some circumstances (or always?), every quest that is inserted into completedRedoableQuests gets the same cooldown (the first entry of completedQuestTimes).
This commit changes the current behaviour, it fetches both lists (completedRedoableQuests/redoableQuestNames and completedQuestTimes/redoableQuestTimes) and matches them together (the first entry of redoableQuestNames gets matched with the first entry of redoableQuestTimes and so on).

If there is not the same amount of entries in both lists, a warning gets printed to the console informing about the possible cooldown-loss.

If there are more entries in redoableQuestNames than in redoableQuestTimes, the missing cooldown a reset with the current timestamp, if there a more entries in redoableQuestTimes than in redoableQuestNames the unnecessary timestamps get deleted and the next save (as they are not in the hashmap).

The warning and the fail-safe-behaviour should never occur but it's better to be safe than sorry.